requirements.txt
pyyaml>=5.3.1
misaka>=2.1.1
jinja2>=2.11.2
Pygments>=2.6.1
chardet>=3.0.4
